# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NSE/BSE AI Trading Signals API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/quotes:
    get:
      operationId: getMarketQuotes
      tags: [market]
      summary: Get live quotes for a list of symbols
      parameters:
        - name: symbols
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated NSE symbols e.g. RELIANCE,NIFTY50,BANKNIFTY
        - name: exchange
          in: query
          schema:
            type: string
            enum: [NSE, BSE]
            default: NSE
      responses:
        "200":
          description: Live quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
...
  /market/options-chain:
    get:
      operationId: getOptionsChain
      tags: [market]
      summary: Get options chain for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
          description: Underlying symbol e.g. NIFTY, BANKNIFTY, RELIANCE
        - name: expiry
          in: query
          schema:
            type: string
          description: Expiry date ISO string; defaults to nearest expiry
      responses:
        "200":
          description: Options chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptionsChain"
  /market/futures:
    get:
      operationId: getFutures
      tags: [market]
      summary: Get futures data for major instruments
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
          description: Symbol filter; omit for all tracked futures
      responses:
        "200":
          description: Futures data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FuturesData"
  /market/history:
    get:
      operationId: getMarketHistory
      tags: [market]
      summary: Get historical OHLCV data for technical analysis
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: interval
          in: query
          schema:
            type: string
            enum: ["1m", "5m", "15m", "1h", "1d"]
            default: "1d"
        - name: period
          in: query
          schema:
            type: string
            enum: ["1d", "5d", "1mo", "3mo", "6mo", "1y"]
            default: "3mo"
      responses:
        "200":
          description: OHLCV history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistory"
...
  /signals:
    get:
      operationId: getSignals
      tags: [signals]
      summary: Get all current AI trading signals
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [STOCK, OPTIONS, FUTURES, ALL]
            default: ALL
        - name: action
          in: query
          schema:
            type: string
            enum: [BUY, SELL, EXIT, ALL]
            default: ALL
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, TRIGGERED, EXPIRED, ALL]
            default: ACTIVE
      responses:
        "200":
          description: Trading signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradingSignal"
  /signals/{id}:
    get:
      operationId: getSignal
      tags: [signals]
      summary: Get a specific signal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingSignal"
        "404":
          description: Not found
  /signals/generate:
    post:
      operationId: generateSignals
      tags: [signals]
      summary: Trigger AI agent to generate new trading signals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSignalsInput"
      responses:
        "200":
          description: Newly generated signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradingSignal"
...
components:
  schemas:
...
    Quote:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        exchange:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        volume:
          type: number
        open:
          type: number
        high:
          type: number
        low:
          type: number
        previousClose:
          type: number
        marketCap:
          type: number
          nullable: true
        timestamp:
          type: string
      required: [symbol, name, exchange, price, change, changePercent, volume, open, high, low, previousClose, timestamp]
...
    OptionsChain:
      type: object
      properties:
        symbol:
          type: string
        underlyingPrice:
          type: number
        expiries:
          type: array
          items:
            type: string
        selectedExpiry:
          type: string
        calls:
          type: array
          items:
            $ref: "#/components/schemas/OptionContract"
        puts:
          type: array
          items:
            $ref: "#/components/schemas/OptionContract"
      required: [symbol, underlyingPrice, expiries, selectedExpiry, calls, puts]
    OptionContract:
      type: object
      properties:
        strikePrice:
          type: number
        expiry:
          type: string
        type:
          type: string
          enum: [CE, PE]
        ltp:
          type: number
        change:
          type: number
        changePercent:
          type: number
        volume:
          type: number
        openInterest:
          type: number
        impliedVolatility:
          type: number
        delta:
          type: number
          nullable: true
        theta:
          type: number
          nullable: true
      required: [strikePrice, expiry, type, ltp, change, changePercent, volume, openInterest, impliedVolatility]
    FuturesData:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        expiry:
          type: string
        ltp:
          type: number
        change:
          type: number
        changePercent:
          type: number
        volume:
          type: number
        openInterest:
          type: number
        basis:
          type: number
      required: [symbol, name, expiry, ltp, change, changePercent, volume, openInterest, basis]
    PriceHistory:
      type: object
      properties:
        symbol:
          type: string
        interval:
          type: string
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
      required: [symbol, interval, candles]
    Candle:
      type: object
      properties:
        timestamp:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      required: [timestamp, open, high, low, close, volume]
...
    TradingSignal:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        instrumentType:
          type: string
          enum: [STOCK, OPTIONS, FUTURES]
        action:
          type: string
          enum: [BUY, SELL, EXIT]
        displayText:
          type: string
          description: Human-readable signal e.g. "BUY NIFTY 22000 CE"
        entryPrice:
          type: number
          nullable: true
        targetPrice:
          type: number
          nullable: true
        stopLoss:
          type: number
          nullable: true
        confidence:
          type: number
          description: 0-100
        rationale:
          type: string
        status:
          type: string
          enum: [ACTIVE, TRIGGERED, EXPIRED]
        timeframe:
          type: string
          enum: [INTRADAY, SWING, POSITIONAL]
        createdAt:
          type: string
        expiresAt:
          type: string
          nullable: true
      required: [id, symbol, instrumentType, action, displayText, confidence, rationale, status, timeframe, createdAt]
    GenerateSignalsInput:
      type: object
      properties:
        symbols:
          type: array
          items:
            type: string
          description: List of symbols to analyze; empty means use watchlist
        timeframe:
          type: string
          enum: [INTRADAY, SWING, POSITIONAL]
          default: INTRADAY
      required: []
...