# lib/api-zod/src/generated/types/optionsChain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NSE/BSE AI Trading Signals API
 * OpenAPI spec version: 0.1.0
 */
import type { OptionContract } from "./optionContract";

export interface OptionsChain {
  symbol: string;
  underlyingPrice: number;
  expiries: string[];
  selectedExpiry: string;
  calls: OptionContract[];
  puts: OptionContract[];
}